+++ src/main/java/org/truffleruby/core/inlined/CoreMethods.java
@@ -7,7 +7,7 @@
  * GNU General Public License version 2
  * GNU Lesser General Public License version 2.1
  */
-package org.truffleruby.core;
+package org.truffleruby.core.inlined;
 
 import com.oracle.truffle.api.object.DynamicObject;
